+++ src/option.ts
@@ -1,17 +1,23 @@
 import { isJust } from './is_just_nothing';
 import { isObject } from './is_object';
 import { isFunction } from './is_function';
-import type { Just } from './utilities';
+import type { Just, Nothing } from './utilities';
 import type { Typeable, Serializable } from './types';
 
 export const OPTION_NONE_OBJECT_TYPE = 'None';
 export const OPTION_SOME_OBJECT_TYPE = 'Some';
 
 export interface Some<T> extends Typeable, Serializable<T> {
-  map<R>(fn: (value: T) => R): Some<R>;
+  map<R>(
+    fn: (value: T) => R
+  ): unknown extends R ? Option<R> : R extends Just<R> ? Some<R> : None;
   fill(): Some<T>;
-  chain<R>(fn: (value: T) => Some<R>): Some<R>;
-  apply<R>(other: Some<(value: T) => R>): Some<R>;
+  chain<R>(
+    fn: (value: T) => Option<R>
+  ): R extends Nothing ? None : unknown extends R ? None : Some<R>;
+  apply<R>(
+    other: Option<(value: T) => R>
+  ): R extends Nothing ? None : unknown extends R ? None : Some<R>;
   isSome(): this is Some<T>;
   isNone(): this is None;
   extract(): T;
@@ -46,8 +52,10 @@ export const none: None = {
 };
 
 export const some = <T>(value: T): Some<T> => ({
-  map: (fn) => some(fn(value)),
+  map: (fn) => maybe(fn(value)),
+  // @ts-ignore
   chain: (fn) => fn(value),
+  // @ts-ignore
   apply: (other) => other.map((fn) => fn(value)),
   fill: () => some(value),
   isNone: () => false,
@@ -64,7 +72,10 @@ export const some = <T>(value: T): Some<T> => ({
  * Depending of the type of _value_ returns `None` or `Some`
  * monad.
  */
-export const maybe = <T>(value: T): Option<Just<T>> =>
+export const maybe = <T>(
+  value: T
+): unknown extends T ? Option<T> : T extends Just<T> ? Some<T> : None =>
+  // @ts-ignore
   isJust(value) ? some(value) : none;
 
 /** Check if value is instance of `Option` monad. */