+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 
 - `Comonad` interface extends now `Functor`.
 - Rename `fold` method of `Foldable` interface to `reduce` due to match [Fantasy Land specification](https://github.com/fantasyland/fantasy-land).
-- Rename `Maybe` type to `Option`, `just` to `some`, `nothing` function to `none` constant.
+- Refactor `Maybe` type to `Option`, `just` to `some`, `nothing` function to `none` constant.
 - Rename `join` method of `List` to `concat`.
 - Rename `join` method of `Stream` to `concat`.
 - `on`, `freeze`, `resume`, `send` and `destroy` methods of `Stream` return now _undefined_ value.